refactor(Mobile): 모바일 컴포넌트에 index 파일 추가
