+++ src/component/Mobile/MobileButton/index.jsx
@@ -1,0 +1,3 @@
+import MobileButton from './MobileButton';
+
+export default MobileButton;
