+++ src/component/Mobile/MobileContact/MobileContact.jsx
@@ -1,5 +1,5 @@
 /** @jsxImportSource @emotion/react */
-import data from '../../../utils/contactData';
+import datas from '../../../utils/contactData';
 import {
   sectionContainer,
   headContainer,
@@ -15,18 +15,18 @@ export default function MobileContact() {
     <section css={sectionContainer}>
       <div css={headContainer}>Contact</div>
       <div css={flexContainer}>
-        {Object.entries(data).map(([key, value]) => (
-          <div key={key} css={infoContainer}>
-            <div css={titleStyle}>{key}</div>
-            {key === 'Github' || key === 'Open Kakao' ? (
-              <div css={infoStyle}>
+        {datas.map(({ title, value, hasLink }) => (
+          <div key={title} css={infoContainer}>
+            <div css={titleStyle}>{title}</div>
+            <div css={infoStyle}>
+              {hasLink ? (
                 <a href={value} target={value} css={linkStyle}>
                   {value}
                 </a>
-              </div>
-            ) : (
-              <div css={infoStyle}>{value}</div>
-            )}
+              ) : (
+                value
+              )}
+            </div>
           </div>
         ))}
       </div>
