+++ src/component/Mobile/MobileContact/index.jsx
@@ -1,0 +1,3 @@
+import MobileContact from './MobileContact';
+
+export default MobileContact;
