+++ src/component/Mobile/MobileContanier/index.jsx
@@ -1,0 +1,3 @@
+import MobileContainer from './MobileContanier';
+
+export default MobileContainer;
