+++ src/component/Mobile/MobileEntertainment/index.jsx
@@ -1,0 +1,3 @@
+import MobileEntertainment from './MobileEntertainment';
+
+export default MobileEntertainment;
