+++ src/component/Mobile/MobileFooter/MobileFooter.jsx
@@ -5,7 +5,7 @@ export default function MobileFooter() {
   return (
     <footer css={sectionContainer}>
       <div css={footerContainer}>
-        <div css={textStyle}>Copyright 2023. Collection50 all right reseved.</div>
+        <div css={textStyle}>Copyright 2023. Collection50 all right reserved.</div>
       </div>
     </footer>
   );
