+++ src/component/Mobile/MobileFooter/index.jsx
@@ -1,0 +1,3 @@
+import MobileFooter from './MobileFooter';
+
+export default MobileFooter;
