+++ src/component/Mobile/MobileHeader/index.jsx
@@ -1,0 +1,3 @@
+import MobileHeader from './MobileHeader';
+
+export default MobileHeader;
