+++ src/component/Mobile/MobileProject/index.jsx
@@ -1,0 +1,3 @@
+import MobileProject from './MobileProject';
+
+export default MobileProject;
