+++ src/component/Mobile/MobilePromotionCard/index.jsx
@@ -1,0 +1,3 @@
+import MobilePromotionCard from './MobilePromotionCard';
+
+export default MobilePromotionCard;
