+++ src/component/Mobile/MobilePromotions/index.jsx
@@ -1,0 +1,3 @@
+import MobilePromotions from './MobilePromotions';
+
+export default MobilePromotions;
