+++ src/component/Mobile/MobileStudy/index.jsx
@@ -1,0 +1,3 @@
+import MobileStudy from './MobileStudy';
+
+export default MobileStudy;
