+++ src/component/Mobile/MobileTimeline/index.jsx
@@ -1,0 +1,3 @@
+import MobileTimeline from './MobileTimeline';
+
+export default MobileTimeline;
