+++ src/component/Mobile/MobileTimelineImage/index.jsx
@@ -1,0 +1,3 @@
+import MobileTimelineImage from './MobileTimelineImage';
+
+export default MobileTimelineImage;
